Conversion to network objects.

Accurately map foreign graph data to network objects.

as_network(x, ...)

# S3 method for bridge_net
as_network(x, .actor = TRUE, ...)

# S3 method for igraph
as_network(x, .actor = TRUE, ...)

# S3 method for network
as_network(x, ...)

# S3 method for tbl_graph
as_network(x, .actor = TRUE, ...)

Arguments

x
bridge_net, igraph (igraph::graph), or tidygraph::tbl_graph.

...
Named arguments passed on to other methods. Currently .actor

.actor
logical, defaults TRUE. Whether to retain the .actor vertex attribute of converted bipartite graphs.

Value

A network::network object.

Details

as_network() converts x to an intermediate bridge_net object that is capable of mapping metadata, edges, vertices, and attributes (edge, vertex, and graph-level) to a new, valid network object.

A note regarding bipartite graph conversion:

network represents bipartite graphs differently than igraph, and thus tidygraph.

network tracks a graph's bipartiteness via the internal indices of each vertex.

Vertices designated as belonging to a graph's "actor" class come before non-"actor" vertices (events, locations, etc.), with "actors" cooresponding to the rows of a graph's affiliation matrix representation.

igraph tracks a graph's bipartiteness via a logical vertex attribute named type, with FALSE and TRUE representing a graph's first and second modes respectively.

The consequence of these very different representations is that converting bipartite igraphs to networks in a way that results in truly identical objects is only possible when an igraph's vertices are already sorted by their `type`` attribute.

To handle this discrepency, intermediate bridge_net objects treat vertices whose type attributes are TRUE as "actors", meaning the igraph's vertices are sorted during conversion to ensure that TRUE vertices are first (via igraph::permute()).

The intention of using TRUE vertices is to allow a user to think of an igraph's vertices' type attribute as indicating whether a vertex is an "actor".

In order to standardize representations, returned bipartite network objects intentionally retain the intermediate bridge_net's .actor vertex attribute (a logical vector).

.actor corresponds to an igraph's type vertex attributes, i.e. V(x_igraph)$type and x_network %v% ".actor" will always match.

If desired, this can be disabled via as_network()'s .actor parameter, i.e. as_network(igraph, .actor = FALSE).

See also

Examples

library(snatools)
library(igraph, warn.conflicts = FALSE, quietly = TRUE)
color_pal <- colorRampPalette(c("red", "green", "blue"))

# simple, 1-mode igraph to network conversion =========================================
ig1 <- random.graph.game(10L, p.or.m = 0.15, directed = TRUE)

V(ig1)$name <- seq_along(V(ig1))
V(ig1)$color <- color_pal(vcount(ig1))
E(ig1)$lab <- seq_along(E(ig1))
ig1

#> IGRAPH 5deb2c5 DN-- 10 14 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n), name (v/n), color
#> | (v/c), lab (e/n)
#> + edges from 5deb2c5 (vertex names):
#>  [1]  5-> 1  4-> 2  7-> 2  2-> 3  1-> 4  4->10  5-> 4 10-> 4 10-> 5  1-> 6
#> [11]  7-> 6  1-> 7  7-> 9 10-> 9

nw1 <- as_network(ig1)
nw1

#>  Network attributes:
#>   vertices = 10 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   name = Erdos renyi (gnp) graph 
#>   type = gnp 
#>   p = 0.15 
#>   total edges= 14 
#>     missing edges= 0 
#>     non-missing edges= 14 
#> 
#>  Vertex attribute names: 
#>     color vertex.names 
#> 
#>  Edge attribute names: 
#>     lab 

# manually comparing original and converted networks ==================================
nw1 %>% vrt_to_df()

#> # vertex_data_frame: 10 x 2
#>      .name   color
#>   1      1 #FF0000
#>   2      2 #C63800
#>   3      3 #8D7100
#>   4      4 #55AA00
#>   5      5 #1CE200
#>   6      6 #00E21C
#>   7      7 #00AA54
#>   8      8 #00718D
#>   9      9 #0038C6
#>   10    10 #0000FF
#> 

ig1 %>% vrt_to_df()

#> # vertex_data_frame: 10 x 2
#>      .name   color
#>   1      1 #FF0000
#>   2      2 #C63800
#>   3      3 #8D7100
#>   4      4 #55AA00
#>   5      5 #1CE200
#>   6      6 #00E21C
#>   7      7 #00AA54
#>   8      8 #00718D
#>   9      9 #0038C6
#>   10    10 #0000FF
#> 

identical(vrt_to_df(nw1), vrt_to_df(ig1))

#> [1] TRUE

nw1 %>% edg_to_df()

#> # edge_data_frame: 14 x 3
#>      .ego .alter lab
#>   1     5      1   1
#>   2     4      2   2
#>   3     7      2   3
#>   4     2      3   4
#>   5     1      4   5
#>   6     4     10   6
#>   7     5      4   7
#>   8    10      4   8
#>   9    10      5   9
#>   10    1      6  10
#>   # ... with 4 additional rows.

ig1 %>% edg_to_df()

#> # edge_data_frame: 14 x 3
#>      .ego .alter lab
#>   1     5      1   1
#>   2     4      2   2
#>   3     7      2   3
#>   4     2      3   4
#>   5     1      4   5
#>   6     4     10   6
#>   7     5      4   7
#>   8    10      4   8
#>   9    10      5   9
#>   10    1      6  10
#>   # ... with 4 additional rows.

identical(edg_to_df(nw1), edg_to_df(ig1))

#> [1] TRUE

# converting bipartite igraph to bipartite network ====================================
ig2 <- bipartite.random.game(3, 5, "gnp", 0.4)
V(ig2)$name <- seq_along(V(ig2))
V(ig2)$color <- ifelse(V(ig2)$type, "red", "blue")
E(ig2)$lab <- sample(seq_len(ecount(ig2)))
ig2

#> IGRAPH 5df5a1d UN-B 8 7 -- Bipartite Gnp random graph
#> + attr: name (g/c), p (g/n), type (v/l), name (v/n), color (v/c), lab
#> | (e/n)
#> + edges from 5df5a1d (vertex names):
#> [1] 2--4 1--6 3--6 1--7 3--7 1--8 2--8

nw2 <- as_network(ig2)
nw2

#>  Network attributes:
#>   vertices = 8 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = 5 
#>   name = Bipartite Gnp random graph 
#>   p = 0.4 
#>   total edges= 7 
#>     missing edges= 0 
#>     non-missing edges= 7 
#> 
#>  Vertex attribute names: 
#>     .actor color vertex.names 
#> 
#>  Edge attribute names: 
#>     lab 

# discarding `.actor` vertex attribute from bipartite graphs ==========================
as_network(ig2, .actor = FALSE)

#>  Network attributes:
#>   vertices = 8 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = 5 
#>   name = Bipartite Gnp random graph 
#>   p = 0.4 
#>   total edges= 7 
#>     missing edges= 0 
#>     non-missing edges= 7 
#> 
#>  Vertex attribute names: 
#>     color vertex.names 
#> 
#>  Edge attribute names: 
#>     lab 

# comparing objects ===================================================================
nw2 %>% vrt_to_df()

#> # vertex_data_frame: 8 x 3
#>     .name .actor color
#>   1     4   TRUE   red
#>   2     5   TRUE   red
#>   3     6   TRUE   red
#>   4     7   TRUE   red
#>   5     8   TRUE   red
#>   6     1  FALSE  blue
#>   7     2  FALSE  blue
#>   8     3  FALSE  blue
#> 

ig2 %>% vrt_to_df()

#> # vertex_data_frame: 8 x 3
#>     .name .actor color
#>   1     4   TRUE   red
#>   2     5   TRUE   red
#>   3     6   TRUE   red
#>   4     7   TRUE   red
#>   5     8   TRUE   red
#>   6     1  FALSE  blue
#>   7     2  FALSE  blue
#>   8     3  FALSE  blue
#> 

identical(vrt_to_df(nw2), vrt_to_df(ig2))

#> [1] TRUE

nw2 %>% edg_to_df()

#> # edge_data_frame: 7 x 3
#>     .ego .alter lab
#>   1    4      2   7
#>   2    6      1   2
#>   3    6      3   1
#>   4    7      1   6
#>   5    7      3   4
#>   6    8      1   5
#>   7    8      2   3
#> 

ig2 %>% edg_to_df()

#> # edge_data_frame: 7 x 3
#>     .ego .alter lab
#>   1    4      2   7
#>   2    6      1   2
#>   3    6      3   1
#>   4    7      1   6
#>   5    7      3   4
#>   6    8      1   5
#>   7    8      2   3
#> 

identical(edg_to_df(nw2), edg_to_df(ig2))

#> [1] TRUE