Davis' Southern Women

Davis' Southern Women

davis_southern_women

Format

network_dataset

Source

http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm

Details

Background, adapted from http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm:

In the 1950s John Gagnon collected sociometric choice data from 67 prison inmates. All were asked, "What fellows on the tier are you closest friends with?" Each was free to choose as few or as many "friends" as he desired. The data were analyzed by MacRae and characterized by him as "less clear cut" in their internal structure than similar data from schools or residential populations.

References

Breiger R. (1974). The duality of persons and groups. Social Forces, 53, 181-190.

Davis, A et al. (1941). Deep South. Chicago: University of Chicago Press.

Examples

data("davis_southern_women", package = "snatools")

davis_southern_women

#> $directed
#> [1] FALSE
#> 
#> $bipartite
#> [1] TRUE
#> 
#> $multiple
#> [1] FALSE
#> 
#> $loops
#> [1] FALSE
#> 
#> $edges
#> # A tibble: 93 x 2
#>    from   to   
#>  * <chr>  <chr>
#>  1 EVELYN E1   
#>  2 EVELYN E2   
#>  3 EVELYN E3   
#>  4 EVELYN E4   
#>  5 EVELYN E5   
#>  6 EVELYN E6   
#>  7 EVELYN E8   
#>  8 EVELYN E9   
#>  9 LAURA  E1   
#> 10 LAURA  E2   
#> # ... with 83 more rows
#> 
#> $vertices
#> # A tibble: 32 x 2
#>    name      type 
#>  * <chr>     <lgl>
#>  1 EVELYN    TRUE 
#>  2 LAURA     TRUE 
#>  3 THERESA   TRUE 
#>  4 BRENDA    TRUE 
#>  5 CHARLOTTE TRUE 
#>  6 FRANCES   TRUE 
#>  7 ELEANOR   TRUE 
#>  8 PEARL     TRUE 
#>  9 RUTH      TRUE 
#> 10 VERNE     TRUE 
#> # ... with 22 more rows
#> 
#> attr(,"class")
#> [1] "network_dataset"

as_igraph(davis_southern_women)

#> IGRAPH 549dc88 UN-B 32 93 -- 
#> + attr: name (v/c), type (v/l)
#> + edges from 549dc88 (vertex names):
#>  [1] EVELYN   --E1 EVELYN   --E2 EVELYN   --E3 EVELYN   --E4 EVELYN   --E5
#>  [6] EVELYN   --E6 EVELYN   --E8 EVELYN   --E9 LAURA    --E1 LAURA    --E2
#> [11] LAURA    --E3 LAURA    --E5 LAURA    --E6 LAURA    --E7 LAURA    --E8
#> [16] THERESA  --E2 THERESA  --E3 THERESA  --E4 THERESA  --E5 THERESA  --E6
#> [21] THERESA  --E7 THERESA  --E8 THERESA  --E9 BRENDA   --E1 BRENDA   --E3
#> [26] BRENDA   --E4 BRENDA   --E5 BRENDA   --E6 BRENDA   --E7 BRENDA   --E8
#> [31] CHARLOTTE--E3 CHARLOTTE--E4 CHARLOTTE--E5 CHARLOTTE--E7 FRANCES  --E3
#> [36] FRANCES  --E5 FRANCES  --E6 FRANCES  --E8 ELEANOR  --E5 ELEANOR  --E6
#> + ... omitted several edges

as_network(davis_southern_women)

#>  Network attributes:
#>   vertices = 32 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   multiple = FALSE 
#>   bipartite = 18 
#>   total edges= 93 
#>     missing edges= 0 
#>     non-missing edges= 93 
#> 
#>  Vertex attribute names: 
#>     is_actor vertex.names 
#> 
#> No edge attributes