Filter edges by their attributes.

Filter edges by their attributes.

edg_filter(x, ...)

edg_filter_se(x, edg_attr, .f, edg_val = NULL)

# S3 method for bridge_net
edg_filter(x, ...)

# S3 method for bridge_net
edg_filter_se(x, edg_attr, .f = NULL,
  edg_val = NULL)

# S3 method for igraph
edg_filter(x, ...)

# S3 method for igraph
edg_filter_se(x, edg_attr, .f = NULL, edg_val = NULL)

# S3 method for network
edg_filter(x, ...)

# S3 method for network
edg_filter_se(x, edg_attr, .f = NULL, edg_val = NULL)

Arguments

x
A bridge_net, igraph, network, or tbl_graph.

...
logical conditions used to subset rows when said conditions evaluate to TRUE.

edg_attr
character specifying edge attribute to target. edg_filter_se() only.

.f
function (default: NULL) specifying the predicate function to apply to edg_attr. edg_filter_se() only.

edg_val
(default: NULL) Value to use as second argument to .f (if applicable).

Value

x, but with edges filtered by ... for edg_filter() or edge_attr, .f, and/or edg_val for edg_filter_se().

Details

edg_filter() uses non-standard evaluation to make syntax as convenient as possible, while edg_filter_se() uses standard evaluation if preferred (or if edg_filter() causes trouble).

Examples

library(snatools)

data(sampson_monastery)

ig <- sampson_monastery %>%
  as_igraph()

nw <- sampson_monastery %>%
  as_network()

sampson_monastery %>%
  edg_filter(relation == "liking", time == 1)

#> # A directed, 1-mode `bridge_net`.
#> # - Contains 0 loops and >0 isolates.
#> $edges # first 3 of 55
#>    .ego    .alter  weight relation time  positive_relation
#>    <.name> <.name> <dbl>  <chr>    <int> <lgl>            
#>  | Romauld Ambrose 2      liking   1     TRUE             
#>  | Romauld Peter   3      liking   1     TRUE             
#>  | Romauld Albert  1      liking   1     TRUE             
#> $vertices # first 3 of 18
#>    .name       faction  cloisterville status  
#>    <.name>     <chr>    <lgl>         <chr>   
#>  | Romauld     Waverers FALSE         Remained
#>  | Bonaventure Loyal     TRUE         Remained
#>  | Ambrose     Loyal    FALSE         Remained

sampson_monastery %>%
  edg_filter_se("relation", `==`, "liking") %>%
  edg_filter_se("time", `==`, 1)

#> # A directed, 1-mode `bridge_net`.
#> # - Contains 0 loops and >0 isolates.
#> $edges # first 3 of 55
#>    .ego    .alter  weight relation time  positive_relation
#>    <.name> <.name> <dbl>  <chr>    <int> <lgl>            
#>  | Romauld Ambrose 2      liking   1     TRUE             
#>  | Romauld Peter   3      liking   1     TRUE             
#>  | Romauld Albert  1      liking   1     TRUE             
#> $vertices # first 3 of 18
#>    .name       faction  cloisterville status  
#>    <.name>     <chr>    <lgl>         <chr>   
#>  | Romauld     Waverers FALSE         Remained
#>  | Bonaventure Loyal     TRUE         Remained
#>  | Ambrose     Loyal    FALSE         Remained

ig %>%
  edg_filter(positive_relation, weight == 3)

#> IGRAPH a1f93a8 DNW- 18 101 -- 
#> + attr: network_name (g/c), author (g/c), name (v/c), faction (v/c),
#> | cloisterville (v/l), status (v/c), weight (e/n), relation (e/c), time
#> | (e/n), positive_relation (e/l)
#> + edges from a1f93a8 (vertex names):
#>  [1] Romauld    ->Peter       Bonaventure->Romauld     Ambrose    ->Bonaventure
#>  [4] Berthold   ->Peter       Peter      ->Berthold    Louis      ->Berthold   
#>  [7] Victor     ->Basil       Winfrid    ->Romauld     John Bosco ->Winfrid    
#> [10] Gregory    ->Berthold    Hugh       ->Peter       Boniface   ->Romauld    
#> [13] Mark       ->Simplicius  Albert     ->Romauld     Amand      ->Romauld    
#> [16] Basil      ->Victor      Elias      ->Ambrose     Simplicius ->Bonaventure
#> + ... omitted several edges

ig %>%
  edg_filter_se("positive_relation") %>%
  edg_filter_se("weight", `==`, 3)

#> IGRAPH a1fc5ef DNW- 18 101 -- 
#> + attr: network_name (g/c), author (g/c), name (v/c), faction (v/c),
#> | cloisterville (v/l), status (v/c), weight (e/n), relation (e/c), time
#> | (e/n), positive_relation (e/l)
#> + edges from a1fc5ef (vertex names):
#>  [1] Romauld    ->Peter       Bonaventure->Romauld     Ambrose    ->Bonaventure
#>  [4] Berthold   ->Peter       Peter      ->Berthold    Louis      ->Berthold   
#>  [7] Victor     ->Basil       Winfrid    ->Romauld     John Bosco ->Winfrid    
#> [10] Gregory    ->Berthold    Hugh       ->Peter       Boniface   ->Romauld    
#> [13] Mark       ->Simplicius  Albert     ->Romauld     Amand      ->Romauld    
#> [16] Basil      ->Victor      Elias      ->Ambrose     Simplicius ->Bonaventure
#> + ... omitted several edges

nw %>%
  edg_filter(relation %in% c("blame", "disesteem"))

#>  Network attributes:
#>   vertices = 18 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = TRUE 
#>   bipartite = FALSE 
#>   network_name = Crisis in the Cloister 
#>   author = Samuel F. Sampson 
#>   total edges= 99 
#>     missing edges= 0 
#>     non-missing edges= 99 
#> 
#>  Vertex attribute names: 
#>     cloisterville faction status vertex.names 
#> 
#>  Edge attribute names: 
#>     positive_relation relation time weight 

nw %>%
  edg_filter_se("relation", `%in%`, c("blame", "disesteem"))

#>  Network attributes:
#>   vertices = 18 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = TRUE 
#>   bipartite = FALSE 
#>   network_name = Crisis in the Cloister 
#>   author = Samuel F. Sampson 
#>   total edges= 99 
#>     missing edges= 0 
#>     non-missing edges= 99 
#> 
#>  Vertex attribute names: 
#>     cloisterville faction status vertex.names 
#> 
#>  Edge attribute names: 
#>     positive_relation relation time weight