Edge Attributes

Extract edge attributes in a variety of formats.

edg_get_attrs(x, ...)

# S3 method for igraph
edg_get_attrs(x)

# S3 method for network
edg_get_attrs(x, ignore_na = TRUE)

edg_get_attrs_df(x, try_tibble = TRUE, ...)

edg_get_attr(x, edg_attr, ...)

# S3 method for network
edg_get_attr(x, edg_attr, ignore_na = TRUE, ...)

edg_get_attr_names(x, ignore_na = TRUE)

Arguments

x
A graph object.

...
Arguments passed on to other methods. See below.

ignore_na
logical. Whether to ignore the "na" attribute of network::network objects.
Default: TRUE

try_tibble
logical. Whether to attempt calling tibble::as_tibble() in edg_get_attrs_df().
Default: TRUE

edg_attr
character. Name of target attribute.

Value

A named list, vector, data.frame, or tibble::tibble. See Details.

Details

edg_get_attrs extract all edge attributes as a named list.

edg_get_attrs_df is a convenience wrapper around edg_get_attrs() that returns a data.frame.

stringsAsFactors is always FALSE.

If available, tibble::as_tibble() is called to return a tbl_df instead.

edg_get_attr extracts the single edge attribute specified by edg_attr as a vector.

edg_get_attr_names extract the names of vertex attributes as a character vector.

See also

Examples

library(snatools)

data("enron", package = "igraphdata")
(ig <- enron)

#> IGRAPH 64ec693 D--- 184 125409 -- Enron email network
#> + attr: LDC_names (g/c), LDC_desc (g/c), name (g/c), Citation (g/c),
#> | Email (v/c), Name (v/c), Note (v/c), Time (e/c), Reciptype (e/c),
#> | Topic (e/n), LDC_topic (e/n)
#> + edges from 64ec693:
#>  [1]  25->154  25->154  30-> 30  30-> 30  30-> 30  30-> 30  39-> 39  52-> 67
#>  [9]  52-> 67  52-> 67  52-> 67  61->100  61->100  61->163  61->163  61->166
#> [17]  61->166  61->170  64-> 59  64-> 59  64-> 64  64-> 64  64->147  64->147
#> [25]  64->164  64->164  64->168  66-> 66  66-> 66  67->129  67->129  67->129
#> [33]  67->129  93-> 10  93-> 10  93-> 10  93-> 10  93-> 39  93-> 39  93-> 93
#> [41]  93-> 93  93-> 93  93-> 93  93->124  93->124 100-> 61 100-> 61 115->115
#> + ... omitted several edges

data("sampson", package = "ergm")
(nw <- samplike)

#>  Network attributes:
#>   vertices = 18 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   total edges= 88 
#>     missing edges= 0 
#>     non-missing edges= 88 
#> 
#>  Vertex attribute names: 
#>     cloisterville group vertex.names 
#> 
#>  Edge attribute names: 
#>     nominations 

edg_get_attr_names(ig)

#> [1] "Time"      "Reciptype" "Topic"     "LDC_topic"

edg_get_attr_names(nw)

#> [1] "nominations"

edg_get_attrs(ig) %>% lapply(head)

#> $Time
#> [1] "1979-12-31 21:00:00" "1979-12-31 21:00:00" "1979-12-31 21:00:00"
#> [4] "1979-12-31 21:00:00" "1979-12-31 21:00:00" "1979-12-31 21:00:00"
#> 
#> $Reciptype
#> [1] "to"  "to"  "cc"  "cc"  "bcc" "bcc"
#> 
#> $Topic
#> [1] 1 1 3 3 3 3
#> 
#> $LDC_topic
#> [1]  0 -1 -1 -1 -1 -1
#> 

edg_get_attrs(nw)

#> $nominations
#>  [1] 1 1 1 1 3 3 2 3 2 3 1 3 3 1 2 2 1 1 3 2 1 1 1 1 2 3 3 3 2 1 3 3 2 1 1 3 1 1
#> [39] 3 2 3 3 1 3 2 3 1 3 2 3 3 1 3 2 3 1 1 3 1 1 1 1 3 1 2 3 1 1 2 1 2 1 1 1 1 1
#> [77] 2 1 1 2 2 2 3 3 1 3 2 3
#> 

edg_get_attrs_df(ig)

#> # A tibble: 125,409 x 6
#>     from    to Time                Reciptype Topic LDC_topic
#>    <dbl> <dbl> <chr>               <chr>     <int>     <int>
#>  1    25   154 1979-12-31 21:00:00 to            1         0
#>  2    25   154 1979-12-31 21:00:00 to            1        -1
#>  3    30    30 1979-12-31 21:00:00 cc            3        -1
#>  4    30    30 1979-12-31 21:00:00 cc            3        -1
#>  5    30    30 1979-12-31 21:00:00 bcc           3        -1
#>  6    30    30 1979-12-31 21:00:00 bcc           3        -1
#>  7    39    39 1979-12-31 21:00:00 to            3        -1
#>  8    52    67 1979-12-31 21:00:00 to            1         0
#>  9    52    67 1979-12-31 21:00:00 to            1         0
#> 10    52    67 1979-12-31 21:00:00 to            3         0
#> # ... with 125,399 more rows

edg_get_attrs_df(nw)

#> # A tibble: 88 x 3
#>    from        to          nominations
#>    <chr>       <chr>             <dbl>
#>  1 Bonaventure John Bosco            1
#>  2 Mark        John Bosco            1
#>  3 John Bosco  Gregory               1
#>  4 Basil       Gregory               1
#>  5 Winfrid     Gregory               3
#>  6 Boniface    Gregory               3
#>  7 John Bosco  Basil                 2
#>  8 Bonaventure Peter                 3
#>  9 John Bosco  Bonaventure           2
#> 10 Peter       Bonaventure           3
#> # ... with 78 more rows

edg_get_attr(ig, edg_attr = "Time") %>% head()

#> [1] "1979-12-31 21:00:00" "1979-12-31 21:00:00" "1979-12-31 21:00:00"
#> [4] "1979-12-31 21:00:00" "1979-12-31 21:00:00" "1979-12-31 21:00:00"

edg_get_attr(nw, edg_attr = "nominations")

#>  [1] 1 1 1 1 3 3 2 3 2 3 1 3 3 1 2 2 1 1 3 2 1 1 1 1 2 3 3 3 2 1 3 3 2 1 1 3 1 1
#> [39] 3 2 3 3 1 3 2 3 1 3 2 3 3 1 3 2 3 1 1 3 1 1 1 1 3 1 2 3 1 1 2 1 2 1 1 1 1 1
#> [77] 2 1 1 2 2 2 3 3 1 3 2 3