Extract Global Network Attributes

Extract global network attributes in a variety of formats consistent across igraph and network objects.

net_get_attrs(x, ...)

# S3 method for igraph
net_get_attrs(x, .default = NA)

# S3 method for network
net_get_attrs(x, drop_metadata = TRUE, .default = NA)

net_get_attrs_df(x, try_tibble = TRUE, ...)

net_get_attr(x, net_attr)

# S3 method for igraph
net_get_attr(x, net_attr, .default = NA)

# S3 method for network
net_get_attr(x, net_attr, .default = NA)

net_get_attr_names(x, ...)

# S3 method for igraph
net_get_attr_names(x, .default = NA)

# S3 method for network
net_get_attr_names(x, drop_metadata = TRUE,
  .default = NA)

Arguments

x
An igraph (igraph::graph) or network::network object.

...
Arguments passed on to other methods. See other arguments.

.default
If returned value is empty (i.e. x has no global network attributes or none that match the target specified by net_attr), .default is returned instead.
Default: NA

drop_metadata
logical. Whether or not to drop a network::network's metadata attributes from returned object. See Details.
Default: TRUE

try_tibble
logical. Whether to attempt calling tibble::as_tibble() in vrt_get_attrs_df().
Default: TRUE

net_attr
character. Name of target attribute.

Value

A named list, vector, data.frame, or tibble::tibble. See Details.

Details

drop_metadata

network::network objects automatically include what can be described as metadata in their global network attributes. These metadata are:

"n" (numeric): The number of vertices in the network object.

"directed" (logical): Whether the network object's edges are meant to be interpreted as directed.

"hyper" (logical): Whether the network object is a hypergraph.

"multiple" (logical): Whether the network object's edges are meant to be interpreted as multiplex.

"bipartite" (logical or numeric): FALSE, if the network object is not bipartite or the number of vertices belonging to the its "actors" mode if it is bipartite.

"mnext" (numeric): The index of the next edge to be added to the network object.

drop_metadata decides whether to include these in the returned object.

net_get_attrs() extracts all global network attributes as a named list.

net_get_attrs_df() is a convenience wrapper around net_get_attrs() that returns a data.frame.

stringsAsFactors is always FALSE.

If try_tibble is TRUE and the tibble package is available, a tibble::tibble is returned.

net_get_attr() extracts the single global, network-level attribute specified by net_attr and returns a scalar.

net_get_attr_names() extracts the names of network/graph-level attributes as a character vector.

Examples

library(snatools)

data("karate", package = "igraphdata")
(ig <- karate)

#> IGRAPH 4b458a1 UNW- 34 78 -- Zachary's karate club network
#> + attr: name (g/c), Citation (g/c), Author (g/c), Faction (v/n), name
#> | (v/c), label (v/c), color (v/n), weight (e/n)
#> + edges from 4b458a1 (vertex names):
#>  [1] Mr Hi  --Actor 2  Mr Hi  --Actor 3  Mr Hi  --Actor 4  Mr Hi  --Actor 5 
#>  [5] Mr Hi  --Actor 6  Mr Hi  --Actor 7  Mr Hi  --Actor 8  Mr Hi  --Actor 9 
#>  [9] Mr Hi  --Actor 11 Mr Hi  --Actor 12 Mr Hi  --Actor 13 Mr Hi  --Actor 14
#> [13] Mr Hi  --Actor 18 Mr Hi  --Actor 20 Mr Hi  --Actor 22 Mr Hi  --Actor 32
#> [17] Actor 2--Actor 3  Actor 2--Actor 4  Actor 2--Actor 8  Actor 2--Actor 14
#> [21] Actor 2--Actor 18 Actor 2--Actor 20 Actor 2--Actor 22 Actor 2--Actor 31
#> [25] Actor 3--Actor 4  Actor 3--Actor 8  Actor 3--Actor 9  Actor 3--Actor 10
#> + ... omitted several edges

data("emon", package = "network")
(nw <- emon$Cheyenne)

#>  Network attributes:
#>   vertices = 14 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   total edges= 83 
#>     missing edges= 0 
#>     non-missing edges= 83 
#> 
#>  Vertex attribute names: 
#>     Command.Rank.Score Decision.Rank.Score Formalization Location Paid.Staff Sponsorship vertex.names Volunteer.Staff 
#> 
#>  Edge attribute names: 
#>     Frequency 

net_get_attrs(ig)

#> $name
#> [1] "Zachary's karate club network"
#> 
#> $Citation
#> [1] "Wayne W. Zachary. An Information Flow Model for Conflict and Fission in Small Groups. Journal of Anthropological Research Vol. 33, No. 4 452-473"
#> 
#> $Author
#> [1] "Wayne W. Zachary"
#> 

net_get_attrs(nw, drop_metadata = FALSE)

#> $n
#> [1] 14
#> 
#> $mnext
#> [1] 84
#> 
#> $directed
#> [1] TRUE
#> 
#> $hyper
#> [1] FALSE
#> 
#> $loops
#> [1] FALSE
#> 
#> $multiple
#> [1] FALSE
#> 

net_get_attrs_df(ig)

#> # A tibble: 1 x 3
#>   name               Citation                                       Author     
#>   <chr>              <chr>                                          <chr>      
#> 1 Zachary's karate ~ Wayne W. Zachary. An Information Flow Model f~ Wayne W. Z~

net_get_attrs_df(nw, drop_metadata = FALSE)

#> # A tibble: 1 x 6
#>       n mnext directed hyper loops multiple
#>   <int> <dbl> <lgl>    <lgl> <lgl> <lgl>   
#> 1    14    84 TRUE     FALSE FALSE FALSE   

net_get_attr(ig, "Author")

#> [1] "Wayne W. Zachary"

net_get_attr_names(ig)

#> [1] "name"     "Citation" "Author"  

net_get_attr_names(nw, drop_metadata = FALSE)

#> [1] "n"        "mnext"    "directed" "hyper"    "loops"    "multiple"