Extract Vertex Attributes

Extract Vertex Attributes

vrt_get_attrs(x, ...)

# S3 method for igraph
vrt_get_attrs(x)

# S3 method for network
vrt_get_attrs(x, ignore_na = TRUE)

vrt_get_attrs_df(x, ...)

vrt_get_attr(x, vrt_attr)

# S3 method for igraph
vrt_get_attr(x, vrt_attr)

# S3 method for network
vrt_get_attr(x, vrt_attr)

vrt_get_names(x)

# S3 method for igraph
vrt_get_names(x)

# S3 method for network
vrt_get_names(x)

vrt_get_attr_names(x, ...)

# S3 method for igraph
vrt_get_attr_names(x)

# S3 method for network
vrt_get_attr_names(x, ignore_na = TRUE)

Arguments

x
A graph object.

...
Arguments passed on to other methods. See below.

ignore_na
logical. Whether to ignore the "na" attribute of network::network objects.
Default: TRUE

vrt_attr
character. Name of target attribute.

try_tibble
logical. Whether to attempt calling tibble::as_tibble() in vrt_get_attrs_df().
Default: TRUE

Value

A named list, vector, data.frame, or tibble::tibble. See Details.

Details

vrt_get_attrs extracts all vertex attributes as a named list.

vrt_get_attrs_df is a convenience wrapper around vrt_get_attrs() that returns a data.frame.

stringsAsFactors is always FALSE.

If available, tibble::as_tibble() is called to return a tbl_df instead.

vrt_get_attr extracts the single vertex attribute specified by vrt_attr as a vector.

vrt_get_names is a convenience wrapper around vrt_get_attr to extract the names of vertices as a vector.

It assumes igraph objects use "name" and network objects use "vertex.names".

If a vertex attribute following this convention is not present, vrt_get_names() throws a warning and returns NULL.

vrt_get_attr_names extract the names of vertex attributes as a character vector.

See also

Examples

library(snatools)

data("karate", package = "igraphdata")
(ig <- karate)

#> IGRAPH 4b458a1 UNW- 34 78 -- Zachary's karate club network
#> + attr: name (g/c), Citation (g/c), Author (g/c), Faction (v/n), name
#> | (v/c), label (v/c), color (v/n), weight (e/n)
#> + edges from 4b458a1 (vertex names):
#>  [1] Mr Hi  --Actor 2  Mr Hi  --Actor 3  Mr Hi  --Actor 4  Mr Hi  --Actor 5 
#>  [5] Mr Hi  --Actor 6  Mr Hi  --Actor 7  Mr Hi  --Actor 8  Mr Hi  --Actor 9 
#>  [9] Mr Hi  --Actor 11 Mr Hi  --Actor 12 Mr Hi  --Actor 13 Mr Hi  --Actor 14
#> [13] Mr Hi  --Actor 18 Mr Hi  --Actor 20 Mr Hi  --Actor 22 Mr Hi  --Actor 32
#> [17] Actor 2--Actor 3  Actor 2--Actor 4  Actor 2--Actor 8  Actor 2--Actor 14
#> [21] Actor 2--Actor 18 Actor 2--Actor 20 Actor 2--Actor 22 Actor 2--Actor 31
#> [25] Actor 3--Actor 4  Actor 3--Actor 8  Actor 3--Actor 9  Actor 3--Actor 10
#> + ... omitted several edges

data("emon", package = "network")
(nw <- emon$Cheyenne)

#>  Network attributes:
#>   vertices = 14 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   total edges= 83 
#>     missing edges= 0 
#>     non-missing edges= 83 
#> 
#>  Vertex attribute names: 
#>     Command.Rank.Score Decision.Rank.Score Formalization Location Paid.Staff Sponsorship vertex.names Volunteer.Staff 
#> 
#>  Edge attribute names: 
#>     Frequency 

vrt_get_attr_names(ig)

#> [1] "Faction" "name"    "label"   "color"  

vrt_get_attr_names(nw)

#> [1] "vertex.names"        "Sponsorship"         "Paid.Staff"         
#> [4] "Volunteer.Staff"     "Location"            "Formalization"      
#> [7] "Decision.Rank.Score" "Command.Rank.Score" 

vrt_get_attrs(ig)

#> $Faction
#>  [1] 1 1 1 1 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 1 2 1 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $name
#>  [1] "Mr Hi"    "Actor 2"  "Actor 3"  "Actor 4"  "Actor 5"  "Actor 6" 
#>  [7] "Actor 7"  "Actor 8"  "Actor 9"  "Actor 10" "Actor 11" "Actor 12"
#> [13] "Actor 13" "Actor 14" "Actor 15" "Actor 16" "Actor 17" "Actor 18"
#> [19] "Actor 19" "Actor 20" "Actor 21" "Actor 22" "Actor 23" "Actor 24"
#> [25] "Actor 25" "Actor 26" "Actor 27" "Actor 28" "Actor 29" "Actor 30"
#> [31] "Actor 31" "Actor 32" "Actor 33" "John A"  
#> 
#> $label
#>  [1] "H"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10" "11" "12" "13" "14" "15"
#> [16] "16" "17" "18" "19" "20" "21" "22" "23" "24" "25" "26" "27" "28" "29" "30"
#> [31] "31" "32" "33" "A" 
#> 
#> $color
#>  [1] 1 1 1 1 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 1 2 1 2 2 2 2 2 2 2 2 2 2 2 2
#> 

vrt_get_attrs(nw)

#> $vertex.names
#>  [1] "Wyoming.Disaster.and.Civil.Defense.Agnecy"     
#>  [2] "Wyoming.State.National.Guard..Army.and.Air."   
#>  [3] "Wyoming.State.Highway.Patrol"                  
#>  [4] "Francis.E..Warren.Air.Force.Base..fire.unit."  
#>  [5] "Red.Cross"                                     
#>  [6] "Salvation.Army"                                
#>  [7] "Laramie.County.Sheriff.s.Office"               
#>  [8] "Laramie.County...Cheyenne.Civil.Defense.Agency"
#>  [9] "Cheyenne.Fire.Department"                      
#> [10] "Cheyenne.Police.Department"                    
#> [11] "Laramie.County.Fire.District..1..volunteer."   
#> [12] "Laramie.County.Fire.District..2..volunteer."   
#> [13] "A.1.Ambulance.Service"                         
#> [14] "Shy.Wy.HAM.Radio.Club"                         
#> 
#> $Sponsorship
#>  [1] "State"       "State"       "State"       "Federal"     "Private"    
#>  [6] "Private"     "County"      "County/City" "City"        "City"       
#> [11] "County"      "County"      "Private"     "Private"    
#> 
#> $Paid.Staff
#>  [1]  10 400 200  60   1   7  60   7  70 100  NA   1  10   0
#> 
#> $Volunteer.Staff
#>  [1]   50 2000    0    0   20   80   20  100    0    0   NA   20    6   NA
#> 
#> $Location
#>  [1] "L" "L" "L" "L" "L" "L" "L" "L" "L" "L" "L" "L" "L" "L"
#> 
#> $Formalization
#>  [1]  2  1  1  1  1  1  1  2  1  3 NA  1  1  1
#> 
#> $Decision.Rank.Score
#>  [1] 20  7  0  5  0  0 20 50 10 20  3  3  2  0
#> 
#> $Command.Rank.Score
#>  [1]  0 10  3  5  0  0 20 40 10 30  2  2  2  0
#> 

vrt_get_attrs_df(ig)

#> # A tibble: 34 x 4
#>    Faction name     label color
#>      <dbl> <chr>    <chr> <dbl>
#>  1       1 Mr Hi    H         1
#>  2       1 Actor 2  2         1
#>  3       1 Actor 3  3         1
#>  4       1 Actor 4  4         1
#>  5       1 Actor 5  5         1
#>  6       1 Actor 6  6         1
#>  7       1 Actor 7  7         1
#>  8       1 Actor 8  8         1
#>  9       2 Actor 9  9         2
#> 10       2 Actor 10 10        2
#> # ... with 24 more rows

vrt_get_attrs_df(nw)

#> # A tibble: 14 x 8
#>    vertex.names Sponsorship Paid.Staff Volunteer.Staff Location Formalization
#>    <chr>        <chr>            <dbl>           <dbl> <chr>            <dbl>
#>  1 Wyoming.Dis~ State               10              50 L                    2
#>  2 Wyoming.Sta~ State              400            2000 L                    1
#>  3 Wyoming.Sta~ State              200               0 L                    1
#>  4 Francis.E..~ Federal             60               0 L                    1
#>  5 Red.Cross    Private              1              20 L                    1
#>  6 Salvation.A~ Private              7              80 L                    1
#>  7 Laramie.Cou~ County              60              20 L                    1
#>  8 Laramie.Cou~ County/City          7             100 L                    2
#>  9 Cheyenne.Fi~ City                70               0 L                    1
#> 10 Cheyenne.Po~ City               100               0 L                    3
#> 11 Laramie.Cou~ County              NA              NA L                   NA
#> 12 Laramie.Cou~ County               1              20 L                    1
#> 13 A.1.Ambulan~ Private             10               6 L                    1
#> 14 Shy.Wy.HAM.~ Private              0              NA L                    1
#> # ... with 2 more variables: Decision.Rank.Score <dbl>,
#> #   Command.Rank.Score <dbl>

vrt_get_attr(ig, vrt_attr = "Faction")

#>  [1] 1 1 1 1 1 1 1 1 2 2 1 1 1 1 2 2 1 1 2 1 2 1 2 2 2 2 2 2 2 2 2 2 2 2

vrt_get_attr(nw, vrt_attr = "Sponsorship")

#>  [1] "State"       "State"       "State"       "Federal"     "Private"    
#>  [6] "Private"     "County"      "County/City" "City"        "City"       
#> [11] "County"      "County"      "Private"     "Private"    

vrt_get_names(ig)

#>  [1] "Mr Hi"    "Actor 2"  "Actor 3"  "Actor 4"  "Actor 5"  "Actor 6" 
#>  [7] "Actor 7"  "Actor 8"  "Actor 9"  "Actor 10" "Actor 11" "Actor 12"
#> [13] "Actor 13" "Actor 14" "Actor 15" "Actor 16" "Actor 17" "Actor 18"
#> [19] "Actor 19" "Actor 20" "Actor 21" "Actor 22" "Actor 23" "Actor 24"
#> [25] "Actor 25" "Actor 26" "Actor 27" "Actor 28" "Actor 29" "Actor 30"
#> [31] "Actor 31" "Actor 32" "Actor 33" "John A"  

vrt_get_names(nw)

#>  [1] "Wyoming.Disaster.and.Civil.Defense.Agnecy"     
#>  [2] "Wyoming.State.National.Guard..Army.and.Air."   
#>  [3] "Wyoming.State.Highway.Patrol"                  
#>  [4] "Francis.E..Warren.Air.Force.Base..fire.unit."  
#>  [5] "Red.Cross"                                     
#>  [6] "Salvation.Army"                                
#>  [7] "Laramie.County.Sheriff.s.Office"               
#>  [8] "Laramie.County...Cheyenne.Civil.Defense.Agency"
#>  [9] "Cheyenne.Fire.Department"                      
#> [10] "Cheyenne.Police.Department"                    
#> [11] "Laramie.County.Fire.District..1..volunteer."   
#> [12] "Laramie.County.Fire.District..2..volunteer."   
#> [13] "A.1.Ambulance.Service"                         
#> [14] "Shy.Wy.HAM.Radio.Club"