Padgett's Florentine Families

florence is an adaptation of the Pajek dataset "Padgett.paj".

florence

Format

Object Class: igraph

directed: FALSE

bipartite: FALSE

multiplex: TRUE

vertices: 16

edges: 35

loops: FALSE

Vertex Attributes:

name

character representing each family's name.

wealth

numeric indicating each family's net wealth in thousands of lira.

priorate_seats indicating the number of seats on the civic council each family held.

Edge Attributes:

relation

character describing relations as either "business" or "marriage".

Source

http://vlado.fmf.uni-lj.si/pub/networks/data/WaFa/default.htm

Details

From the Pajek documentation (originally from the UCINET IV documentation):

Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data on the social relations among Renaissance Florentine families (person aggregates) collected by John Padgett from historical documents. The two relations are business ties (PADGB - specifically, recorded financial ties such as loans, credits and joint partnerships) and marriage alliances (PADGM).

As Breiger & Pattison point out, the original data are symmetrically coded. This is acceptable perhaps for marital ties, but is unfortunate for the financial ties (which are almost certainly directed). To remedy this, the financial ties can be recoded as directed relations using some external measure of power - for instance, a measure of wealth. PADGW provides information on (1) each family's net wealth in 1427 (in thousands of lira); (2) the number of priorates (seats on the civic council) held between 1282- 1344; and (3) the total number of business or marriage ties in the total dataset of 116 families (see Breiger & Pattison (1986), p 239).

Substantively, the data include families who were locked in a struggle for political control of the city of Florence in around 1430. Two factions were dominant in this struggle: one revolved around the infamous Medicis (9), the other around the powerful Strozzis (15).

References

Breiger R. and Pattison P. (1986). Cumulated social roles: The duality of persons and their algebras. Social Networks, 8, 215-256.

Kent D. (1978). The rise of the Medici: Faction in Florence, 1426-1434. Oxford: Oxford University Press.

Examples

library(snatools)

florence

#> IGRAPH bd28944 UN-- 16 35 -- 
#> + attr: .vrt_id (v/n), name (v/c), wealth (v/n), priorate_seats (v/n),
#> | .edg_id (e/n), relation (e/c)
#> + edges from bd28944 (vertex names):
#>  [1] Barbadori   --Castellani   Barbadori   --Ginori      
#>  [3] Barbadori   --Medici       Barbadori   --Peruzzi     
#>  [5] Bischeri    --Guadagni     Bischeri    --Lamberteschi
#>  [7] Bischeri    --Peruzzi      Castellani  --Lamberteschi
#>  [9] Castellani  --Peruzzi      Ginori      --Medici      
#> [11] Guadagni    --Lamberteschi Lamberteschi--Peruzzi     
#> [13] Medici      --Pazzi        Medici      --Salviati    
#> + ... omitted several edges

as_network(florence)

#>  Network attributes:
#>   vertices = 16 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = TRUE 
#>   bipartite = FALSE 
#>   total edges= 35 
#>     missing edges= 0 
#>     non-missing edges= 35 
#> 
#>  Vertex attribute names: 
#>     .vrt_id priorate_seats vertex.names wealth 
#> 
#>  Edge attribute names: 
#>     .edg_id relation