Metadata Queries

Metadata Queries

net_is_bipartite(x)

# S3 method for igraph
net_is_bipartite(x)

# S3 method for network
net_is_bipartite(x)

net_is_directed(x)

# S3 method for igraph
net_is_directed(x)

# S3 method for network
net_is_directed(x)

Arguments

x
A graph object.

Details

net_is_bipartite() checks whether x is labeled as bipartite and returns a logical scalar.

net_is_directed() checks whether x is labeled as directed and returns a logical scalar.

Examples

library(snatools)

data("karate", package = "igraphdata")
(ig <- karate)

#> IGRAPH 4b458a1 UNW- 34 78 -- Zachary's karate club network
#> + attr: name (g/c), Citation (g/c), Author (g/c), Faction (v/n), name
#> | (v/c), label (v/c), color (v/n), weight (e/n)
#> + edges from 4b458a1 (vertex names):
#>  [1] Mr Hi  --Actor 2  Mr Hi  --Actor 3  Mr Hi  --Actor 4  Mr Hi  --Actor 5 
#>  [5] Mr Hi  --Actor 6  Mr Hi  --Actor 7  Mr Hi  --Actor 8  Mr Hi  --Actor 9 
#>  [9] Mr Hi  --Actor 11 Mr Hi  --Actor 12 Mr Hi  --Actor 13 Mr Hi  --Actor 14
#> [13] Mr Hi  --Actor 18 Mr Hi  --Actor 20 Mr Hi  --Actor 22 Mr Hi  --Actor 32
#> [17] Actor 2--Actor 3  Actor 2--Actor 4  Actor 2--Actor 8  Actor 2--Actor 14
#> [21] Actor 2--Actor 18 Actor 2--Actor 20 Actor 2--Actor 22 Actor 2--Actor 31
#> [25] Actor 3--Actor 4  Actor 3--Actor 8  Actor 3--Actor 9  Actor 3--Actor 10
#> + ... omitted several edges

data("emon", package = "network")
(nw <- emon$Cheyenne)

#>  Network attributes:
#>   vertices = 14 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   total edges= 83 
#>     missing edges= 0 
#>     non-missing edges= 83 
#> 
#>  Vertex attribute names: 
#>     Command.Rank.Score Decision.Rank.Score Formalization Location Paid.Staff Sponsorship vertex.names Volunteer.Staff 
#> 
#>  Edge attribute names: 
#>     Frequency 

net_is_bipartite(ig)

#> [1] FALSE

net_is_bipartite(nw)

#> [1] FALSE

net_is_directed(ig)

#> [1] FALSE

net_is_directed(nw)

#> [1] TRUE