Padgett's Florentine Families

Padgett's Florentine Families

padgett_florence

Format

network_dataset

Source

http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm

Details

Background, adapted from http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm:

Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data on the social relations among Renaissance Florentine families (person aggregates) collected by John Padgett from historical documents. The two relations are business ties (specifically, recorded financial ties such as loans, credits and joint partnerships) and marriage alliances.

As Breiger & Pattison point out, the original data are symmetrically coded. This is acceptable perhaps for marital ties, but is unfortunate for the financial ties (which are almost certainly directed). To remedy this, the financial ties can be recoded as directed relations using some external measure of power - for instance, a measure of wealth. PADGW provides information on (1) each family's net wealth in 1427 (in thousands of lira); (2) the number of priorates (seats on the civic council) held between 1282- 1344; and (3) the total number of business or marriage ties in the total dataset of 116 families (see Breiger & Pattison (1986), p 239).

Subtantively, the data include families who were locked in a struggle for political control of the city of Florence in around 1430. Two factions were dominant in this struggle: one revolved around the infamous Medicis (9), the other around the powerful Strozzis (15).

References

Breiger R. and Pattison P. (1986). Cumulated social roles: The duality of persons and their algebras. Social Networks, 8, 215-256.

Kent D. (1978). The rise of the Medici: Faction in Florence, 1426-1434. Oxford: Oxford University Press.

Examples

data("padgett_florence", package = "snatools")

padgett_florence

#> $graph_attributes
#> NULL
#> 
#> $metadata
#> $metadata$n_vertices
#> [1] 16
#> 
#> $metadata$n_edges
#> [1] 35
#> 
#> $metadata$is_directed
#> [1] FALSE
#> 
#> $metadata$is_bipartite
#> [1] FALSE
#> 
#> $metadata$any_multiplex
#> [1] TRUE
#> 
#> $metadata$any_loops
#> [1] FALSE
#> 
#> 
#> $edgelist
#>    ego alter
#> 1    1     9
#> 2    2     6
#> 3    2     7
#> 4    2     9
#> 5    3     5
#> 6    3     9
#> 7    4     7
#> 8    4    11
#> 9    4    15
#> 10   5    11
#> 11   5    15
#> 12   7     8
#> 13   7    16
#> 14   9    13
#> 15   9    14
#> 16   9    16
#> 17  10    14
#> 18  11    15
#> 19  13    15
#> 20  13    16
#> 21   3     5
#> 22   3     6
#> 23   3     9
#> 24   3    11
#> 25   4     7
#> 26   4     8
#> 27   4    11
#> 28   5     8
#> 29   5    11
#> 30   6     9
#> 31   7     8
#> 32   8    11
#> 33   9    10
#> 34   9    14
#> 35   9    16
#> 
#> $edge_attributes
#>    relation
#> 1  marriage
#> 2  marriage
#> 3  marriage
#> 4  marriage
#> 5  marriage
#> 6  marriage
#> 7  marriage
#> 8  marriage
#> 9  marriage
#> 10 marriage
#> 11 marriage
#> 12 marriage
#> 13 marriage
#> 14 marriage
#> 15 marriage
#> 16 marriage
#> 17 marriage
#> 18 marriage
#> 19 marriage
#> 20 marriage
#> 21 business
#> 22 business
#> 23 business
#> 24 business
#> 25 business
#> 26 business
#> 27 business
#> 28 business
#> 29 business
#> 30 business
#> 31 business
#> 32 business
#> 33 business
#> 34 business
#> 35 business
#> 
#> $vertex_attributes
#>         name net_wealth priorates
#> 1  Acciaiuol         10        53
#> 2    Albizzi         36        65
#> 3  Barbadori         55         0
#> 4   Bischeri         44        12
#> 5  Castellan         20        22
#> 6     Ginori         32         0
#> 7   Guadagni          8        21
#> 8  Lambertes         42         0
#> 9     Medici        103        53
#> 10     Pazzi         48         0
#> 11   Peruzzi         49        42
#> 12     Pucci          3         0
#> 13   Ridolfi         27        38
#> 14  Salviati         10        35
#> 15   Strozzi        146        74
#> 16 Tornabuon         48         0
#> 
#> attr(,"class")
#> [1] "net_primitive"

as_igraph(padgett_florence)

#> Error in UseMethod("as_igraph"): no applicable method for 'as_igraph' applied to an object of class "net_primitive"

as_network(padgett_florence)

#> Error in UseMethod("as_network"): no applicable method for 'as_network' applied to an object of class "net_primitive"