Construct a graph object's adjacency list representation.

Construct a graph object's adjacency list representation.

rep_as_adj_list(x, use_names = TRUE, vrt_attr = NULL)

Arguments

x
An igraph, network, or tbl_graph.

use_names
logical (default: TRUE) indicating whether to use vertex names as the values filling the returned object. If FALSE, or a relevant attribute does not exist, the returned object uses vertex indices.

vrt_attr
character specifying the target vertex attribute.

Value

list, where names and values correspond to vertex names, indices, or attributes, depending on arguments provided.

See also

Examples

library(snatools)

jemmah_islamiyah %>%
  rep_as_adj_list()

#> $MUKLAS
#> [1] "AMROZI"   "IMRON"    "SAMUDRA"  "DULMATIN" "IDRIS"    "AZAHARI"  "GHONI"   
#> [8] "PATEK"    "SARIJO"  
#> 
#> $AMROZI
#> [1] "SAMUDRA" "IDRIS"   "MUBAROK" "MUKLAS" 
#> 
#> $IMRON
#> [1] "SAMUDRA"  "DULMATIN" "IDRIS"    "AZAHARI"  "GHONI"    "PATEK"    "FERI"    
#> [8] "SARIJO"   "MUKLAS"  
#> 
#> $SAMUDRA
#>  [1] "DULMATIN" "IDRIS"    "MUBAROK"  "AZAHARI"  "GHONI"    "ARNASAN" 
#>  [7] "RAUF"     "OCTAVIA"  "HIDAYAT"  "JUNAEDI"  "PATEK"    "SARIJO"  
#> [13] "MUKLAS"   "AMROZI"   "IMRON"   
#> 
#> $DULMATIN
#> [1] "IDRIS"   "AZAHARI" "GHONI"   "PATEK"   "FERI"    "SARIJO"  "MUKLAS" 
#> [8] "IMRON"   "SAMUDRA"
#> 
#> $IDRIS
#>  [1] "MUBAROK"  "AZAHARI"  "GHONI"    "PATEK"    "SARIJO"   "MUKLAS"  
#>  [7] "AMROZI"   "IMRON"    "SAMUDRA"  "DULMATIN"
#> 
#> $MUBAROK
#> [1] "AMROZI"  "SAMUDRA" "IDRIS"  
#> 
#> $AZAHARI
#> [1] "GHONI"    "PATEK"    "FERI"     "SARIJO"   "MUKLAS"   "IMRON"    "SAMUDRA" 
#> [8] "DULMATIN" "IDRIS"   
#> 
#> $GHONI
#> [1] "PATEK"    "FERI"     "SARIJO"   "MUKLAS"   "IMRON"    "SAMUDRA"  "DULMATIN"
#> [8] "IDRIS"    "AZAHARI" 
#> 
#> $ARNASAN
#> [1] "RAUF"    "OCTAVIA" "HIDAYAT" "JUNAEDI" "SAMUDRA"
#> 
#> $RAUF
#> [1] "OCTAVIA" "HIDAYAT" "JUNAEDI" "SAMUDRA" "ARNASAN"
#> 
#> $OCTAVIA
#> [1] "HIDAYAT" "JUNAEDI" "SAMUDRA" "ARNASAN" "RAUF"   
#> 
#> $HIDAYAT
#> [1] "JUNAEDI" "SAMUDRA" "ARNASAN" "RAUF"    "OCTAVIA"
#> 
#> $JUNAEDI
#> [1] "SAMUDRA" "ARNASAN" "RAUF"    "OCTAVIA" "HIDAYAT"
#> 
#> $PATEK
#> [1] "FERI"     "SARIJO"   "MUKLAS"   "IMRON"    "SAMUDRA"  "DULMATIN" "IDRIS"   
#> [8] "AZAHARI"  "GHONI"   
#> 
#> $FERI
#> [1] "SARIJO"   "IMRON"    "DULMATIN" "AZAHARI"  "GHONI"    "PATEK"   
#> 
#> $SARIJO
#> [1] "MUKLAS"   "IMRON"    "SAMUDRA"  "DULMATIN" "IDRIS"    "AZAHARI"  "GHONI"   
#> [8] "PATEK"    "FERI"    
#> 

jemmah_islamiyah %>%
  rep_as_adj_list(use_names = FALSE)

#> [[1]]
#> [1]  2  3  4  5  6  8  9 15 17
#> 
#> [[2]]
#> [1] 4 6 7 1
#> 
#> [[3]]
#> [1]  4  5  6  8  9 15 16 17  1
#> 
#> [[4]]
#>  [1]  5  6  7  8  9 10 11 12 13 14 15 17  1  2  3
#> 
#> [[5]]
#> [1]  6  8  9 15 16 17  1  3  4
#> 
#> [[6]]
#>  [1]  7  8  9 15 17  1  2  3  4  5
#> 
#> [[7]]
#> [1] 2 4 6
#> 
#> [[8]]
#> [1]  9 15 16 17  1  3  4  5  6
#> 
#> [[9]]
#> [1] 15 16 17  1  3  4  5  6  8
#> 
#> [[10]]
#> [1] 11 12 13 14  4
#> 
#> [[11]]
#> [1] 12 13 14  4 10
#> 
#> [[12]]
#> [1] 13 14  4 10 11
#> 
#> [[13]]
#> [1] 14  4 10 11 12
#> 
#> [[14]]
#> [1]  4 10 11 12 13
#> 
#> [[15]]
#> [1] 16 17  1  3  4  5  6  8  9
#> 
#> [[16]]
#> [1] 17  3  5  8  9 15
#> 
#> [[17]]
#> [1]  1  3  4  5  6  8  9 15 16
#>