Construct a graph object's edgelist representation.

Construct a graph object's edgelist representation.

rep_as_adj_matrix(x, use_names = TRUE, vrt_attr = NULL,
  edg_attr = NULL)

Arguments

x
An igraph, network, or tbl_graph.

use_names
logical (default: TRUE) indicating whether to use vertex names as the values filling the returned object. If FALSE, or a relevant attribute does not exist, the returned object uses vertex indices.

vrt_attr
character specifying the target vertex attribute.

edg_attr
character (default: NULL) indicating which edge attribute to use as fill values in the returned object.

Value

matrix with net_count_vertices(x) rows and columns

See also

Examples

library(snatools)

florence %>%
  edg_subset(relation == "marriage") %>%
  rep_as_adj_matrix() %>%
  `colnames<-`(rep("", ncol(.)))

#>                                             
#> Acciaiuoli   0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> Albizzi      0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 0
#> Barbadori    0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0
#> Bischeri     0 0 0 0 0 0 1 0 0 0 1 0 0 0 1 0
#> Castellani   0 0 0 0 0 0 0 0 0 0 1 0 0 0 1 0
#> Ginori       0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Guadagni     0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 1
#> Lamberteschi 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Medici       0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 1
#> Pazzi        0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> Peruzzi      0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> Pucci        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Ridolfi      0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1
#> Salviati     0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Strozzi      0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Tornabuoni   0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0

crisis_in_cloister %>%
  edg_subset(time == 1 & affect == 3) %>%
  rep_as_adj_matrix() %>%
  `colnames<-`(rep("", ncol(.)))

#>                                                              
#> Leo         0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> Arsenius    0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Bruno       0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Thomas      0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> Bartholomew 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> John Bosco  0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Gregory     0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Basil       0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Martin      0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Peter       0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Bonaventure 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Berthold    0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Mark        1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Brocard     1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Victor      0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Ambrose     0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> Ramuald     0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Louis       0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Winfrid     0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Amand       0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Hugh        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Boniface    0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Albert      0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Elias       0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> Simplicius  0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0

florence %>%
  rep_as_adj_matrix(edg_attr = "relation") %>%
  `colnames<-`(rep("", ncol(.)))

#>                                                                                
#> Acciaiuoli   NA NA NA NA NA         NA         NA         NA         "marriage"
#> Albizzi      NA NA NA NA NA         "marriage" "marriage" NA         "marriage"
#> Barbadori    NA NA NA NA "marriage" "business" NA         NA         "marriage"
#> Bischeri     NA NA NA NA NA         NA         "marriage" "business" NA        
#> Castellani   NA NA NA NA NA         NA         NA         "business" NA        
#> Ginori       NA NA NA NA NA         NA         NA         NA         "business"
#> Guadagni     NA NA NA NA NA         NA         NA         "marriage" NA        
#> Lamberteschi NA NA NA NA NA         NA         NA         NA         NA        
#> Medici       NA NA NA NA NA         NA         NA         NA         NA        
#> Pazzi        NA NA NA NA NA         NA         NA         NA         NA        
#> Peruzzi      NA NA NA NA NA         NA         NA         NA         NA        
#> Pucci        NA NA NA NA NA         NA         NA         NA         NA        
#> Ridolfi      NA NA NA NA NA         NA         NA         NA         NA        
#> Salviati     NA NA NA NA NA         NA         NA         NA         NA        
#> Strozzi      NA NA NA NA NA         NA         NA         NA         NA        
#> Tornabuoni   NA NA NA NA NA         NA         NA         NA         NA        
#>                                                                       
#> Acciaiuoli   NA         NA         NA NA         NA         NA        
#> Albizzi      NA         NA         NA NA         NA         NA        
#> Barbadori    NA         "business" NA NA         NA         NA        
#> Bischeri     NA         "marriage" NA NA         NA         "marriage"
#> Castellani   NA         "marriage" NA NA         NA         "marriage"
#> Ginori       NA         NA         NA NA         NA         NA        
#> Guadagni     NA         NA         NA NA         NA         NA        
#> Lamberteschi NA         "business" NA NA         NA         NA        
#> Medici       "business" NA         NA "marriage" "marriage" NA        
#> Pazzi        NA         NA         NA NA         "marriage" NA        
#> Peruzzi      NA         NA         NA NA         NA         "marriage"
#> Pucci        NA         NA         NA NA         NA         NA        
#> Ridolfi      NA         NA         NA NA         NA         "marriage"
#> Salviati     NA         NA         NA NA         NA         NA        
#> Strozzi      NA         NA         NA NA         NA         NA        
#> Tornabuoni   NA         NA         NA NA         NA         NA        
#>                        
#> Acciaiuoli   NA        
#> Albizzi      NA        
#> Barbadori    NA        
#> Bischeri     NA        
#> Castellani   NA        
#> Ginori       NA        
#> Guadagni     "marriage"
#> Lamberteschi NA        
#> Medici       "marriage"
#> Pazzi        NA        
#> Peruzzi      NA        
#> Pucci        NA        
#> Ridolfi      "marriage"
#> Salviati     NA        
#> Strozzi      NA        
#> Tornabuoni   NA