Construct a graph object's edgelist representation

rep_as_edgelist() standardizes edgelist representations across supported classes in that edge sources (.ego) and targets (.alter) are sorted row-wise for undirected graphs and parallel edges are never dropped from multiplex graphs.

rep_as_edgelist(x, use_names = TRUE, vrt_attr = NULL,
  weights = FALSE, leave_raw = FALSE, ignore_missing_names = FALSE)

is_edgelist(x)

# S3 method for edgelist
print(x, .nrow = 10L)

# S3 method for edgelist
as.matrix(x, ...)

# S3 method for edgelist
as.data.frame(x, ...)

Arguments

x
An bridge_net, igraph, network, tbl_graph, or edgelist object.

use_names
logical (default: TRUE) indicating whether to use vertex names in the returned object. If FALSE, the returned object uses vertex indices.

vrt_attr
character (default: NULL) indicating which vertex attribute to use in the returned object. If provided, vrt_attr overrides use_names.

weights
logical (default: FALSE) indicating whether the parallel edges of multiplex graph objects should be combined into a count column named .weight.

leave_raw
logical (default: TRUE) indicating whether to return an edgelist object instead of a matrix or data.frame.

ignore_missing_names
logical (default: FALSE) whether to throw a warning if use_names = TRUE but x does not contain valid vertex names.

.nrow
Maximum number of rows to print.

...
Arguments passed on to as.data.frame().

Value

edgelist object, matrix, or data.frame.

Details

The returned object contains minimally 2 columns (.ego and .alter) stored as a matrix.

use_names

If use_names is TRUE (the default), rep_as_edgelist() checks whether x contains a valid vertex attribute for vertex names. If available, this attribute is used. If not available, the returned object uses vertex indices and a warning is thrown (unless ignore_missing_names = TRUE).

Each class uses a different attribute for valid vertex names:

igraph and tbl_graph objects use a vertex attribute called name

network objects use a vertex attribute called vertex.names

bridge_net objects use a vertex attribute called .name

weights

If weights is TRUE, rep_as_edgelist() combines parallel edges to create a third column (.weight).

In order to handle potentially heterogeneous data types, rep_as_edgelist() stores edges in a data.frame instead of a matrix.

Ignored if net_is_multiplex() returns FALSE.

leave_raw

If leave_raw is FALSE (the default), rep_as_edgelist() will return an edgelist object.

edgelist objects store metadata describing x in order to facilitate operations for which a traditional edgelist requires extra information.

These metadata are stored in the following attributes():

is_directed

is_bipartite

is_multiplex

n_vertices

n_edges

n_actors

If desired, an edgelist object can be converted using as.data.frame() or as.matrix() in order to strip additional attributes().

See also

Examples

library(snatools)

ig <- igraph::random.graph.game(10, 0.1) %>%
  igraph::set_vertex_attr("name", value = letters[seq_len(igraph::vcount(.))]) %>%
  igraph::set_vertex_attr("group", value = sample(LETTERS[1:3], igraph::vcount(.), replace = TRUE))

nw <- ig %>%
  as_network()

sna_g <- ig %>%
  as_bridge_net()

tidy_g <- ig %>%
  tidygraph::as_tbl_graph()

ig %>% rep_as_edgelist()

#> # An undirected edgelist with 4 edges.*
#>   .ego .alter
#> 1 b    d     
#> 2 d    e     
#> 3 b    g     
#> 4 h    j     
#> # * Values correspond to vertex names.

nw %>% rep_as_edgelist(use_names = FALSE)

#> # An undirected edgelist with 4 edges.*
#>   .ego .alter
#> 1    2      4
#> 2    4      5
#> 3    2      7
#> 4    8     10
#> # * Values correspond to vertex indices.

tidy_g %>% rep_as_edgelist(vrt_attr = "group")

#> # An undirected edgelist with 4 edges.*
#>   .ego .alter
#> 1 B    B     
#> 2 B    A     
#> 3 B    B     
#> 4 A    A     
#> # * Values correspond to a vertex attribute ('group').

ig %>% rep_as_edgelist(leave_raw = TRUE)

#>      .ego .alter
#> [1,] "b"  "d"   
#> [2,] "d"  "e"   
#> [3,] "b"  "g"   
#> [4,] "h"  "j"