Sampson's Monastery

Sampson's Monastery

sampson_monastery

Format

network_dataset

Source

http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm

Details

Background, adapted from http://vlado.fmf.uni-lj.si/pub/networks/data/ucinet/ucidata.htm:

Sampson recorded the social interactions among a group of monks while resident as an experimenter on vision, and collected numerous sociometric rankings. During his stay, a political "crisis in the cloister" resulted in the expulsion of four monks (Nos. 2, 3, 17, and 18) and the voluntary departure of several others - most immediately, Nos. 1, 7, 14, 15, and 16. (In the end, only 5, 6, 9, and 11 remained).

Most of the present data are retrospective, collected after the breakup occurred. They concern a period during which a new cohort entered the monastery near the end of the study but before the major conflict began. The exceptions are "liking" data gathered at three times: SAMPLK1 to SAMPLK3 - that reflect changes in group sentiment over time (SAMPLK3 was collected in the same wave as the data described below). Information about the senior monks was not included.

Four relations are coded, with separate matrices for positive and negative ties on the relation. Each member ranked only his top three choices on that tie. The relations are esteem (SAMPES) and disesteem (SAMPDES), liking (SAMPLK) and disliking (SAMPDLK), positive influence (SAMPIN) and negative influence (SAMPNIN), praise (SAMPPR) and blame (SAMPNPR). In all rankings 3 indicates the highest or first choice and 1 the last choice. (Some subjects offered tied ranks for their top four choices).

References

Breiger R., Boorman S. and Arabie P. (1975). An algorithm for clustering relational data with applications to social network analysis and comparison with multidimensional scaling. Journal of Mathematical Psychology, 12, 328-383.

Sampson, S. (1969). Crisis in a cloister. Unpublished doctoral dissertation, Cornell University.

Examples

data("sampson_monastery", package = "snatools")

sampson_monastery

#> $directed
#> [1] TRUE
#> 
#> $bipartite
#> [1] FALSE
#> 
#> $multiple
#> [1] TRUE
#> 
#> $loops
#> [1] FALSE
#> 
#> $edges
#> # A tibble: 510 x 6
#>    from        to          weight relation  time positive_relation
#>    <chr>       <chr>        <dbl> <chr>    <int> <lgl>            
#>  1 Romauld     Ambrose          2 liking       1 TRUE             
#>  2 Romauld     Peter            3 liking       1 TRUE             
#>  3 Romauld     Albert           1 liking       1 TRUE             
#>  4 Bonaventure Romauld          3 liking       1 TRUE             
#>  5 Bonaventure Victor           2 liking       1 TRUE             
#>  6 Bonaventure Albert           1 liking       1 TRUE             
#>  7 Ambrose     Romauld          2 liking       1 TRUE             
#>  8 Ambrose     Bonaventure      3 liking       1 TRUE             
#>  9 Ambrose     Elias            1 liking       1 TRUE             
#> 10 Berthold    Peter            3 liking       1 TRUE             
#> # ... with 500 more rows
#> 
#> $vertices
#> # A tibble: 18 x 4
#>    name        faction     cloisterville status          
#>    <chr>       <chr>       <lgl>         <chr>           
#>  1 Romauld     Waverers    FALSE         Remained        
#>  2 Bonaventure Loyal       TRUE          Remained        
#>  3 Ambrose     Loyal       FALSE         Remained        
#>  4 Berthold    Loyal       FALSE         Remained        
#>  5 Peter       Loyal       TRUE          Remained        
#>  6 Louis       Loyal       FALSE         Remained        
#>  7 Victor      Waverers    FALSE         Remained        
#>  8 Winfrid     Young Turks FALSE         Remained        
#>  9 John Bosco  Young Turks TRUE          Left Voluntarily
#> 10 Gregory     Young Turks TRUE          Expelled        
#> 11 Hugh        Young Turks FALSE         Left Voluntarily
#> 12 Boniface    Young Turks FALSE         Left Voluntarily
#> 13 Mark        Young Turks TRUE          Left Voluntarily
#> 14 Albert      Young Turks FALSE         Left Voluntarily
#> 15 Amand       Waverers    FALSE         Remained        
#> 16 Basil       Outcasts    TRUE          Expelled        
#> 17 Elias       Outcasts    FALSE         Expelled        
#> 18 Simplicius  Outcasts    FALSE         Expelled        
#> 
#> attr(,"class")
#> [1] "network_dataset"

as_igraph(sampson_monastery)

#> IGRAPH 450655e DNW- 18 510 -- 
#> + attr: name (v/c), faction (v/c), cloisterville (v/l), status (v/c),
#> | weight (e/n), relation (e/c), time (e/n), positive_relation (e/l)
#> + edges from 450655e (vertex names):
#>  [1] Romauld    ->Ambrose     Romauld    ->Peter       Romauld    ->Albert     
#>  [4] Bonaventure->Romauld     Bonaventure->Victor      Bonaventure->Albert     
#>  [7] Ambrose    ->Romauld     Ambrose    ->Bonaventure Ambrose    ->Elias      
#> [10] Berthold   ->Peter       Berthold   ->Louis       Berthold   ->Gregory    
#> [13] Peter      ->Berthold    Peter      ->Hugh        Peter      ->Mark       
#> [16] Louis      ->Romauld     Louis      ->Berthold    Louis      ->John Bosco 
#> [19] Victor     ->Bonaventure Victor     ->Winfrid     Victor     ->Basil      
#> + ... omitted several edges

as_network(sampson_monastery)

#>  Network attributes:
#>   vertices = 18 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   multiple = TRUE 
#>   total edges= 510 
#>     missing edges= 0 
#>     non-missing edges= 510 
#> 
#>  Vertex attribute names: 
#>     cloisterville faction status vertex.names 
#> 
#>  Edge attribute names: 
#>     positive_relation relation time weight