The sna_net object effectively bridges the network objects used by the statnet suite and igraph objects and their derivatives.

The sna_net object effectively bridges the network objects used by the statnet suite and igraph objects and their derivatives.

sna_net(edges = NULL, vertices = NULL, is_directed = FALSE)

as_sna_net(x, ...)

# S3 method for sna_net
as_sna_net(x)

# S3 method for network
as_sna_net(x, edg_attrs = TRUE, vrt_attrs = TRUE,
  net_attrs = TRUE)

# S3 method for igraph
as_sna_net(x, edg_attrs = TRUE, vrt_attrs = TRUE,
  net_attrs = TRUE)

# S3 method for tbl_graph
as_sna_net(x, edg_attrs = TRUE, vrt_attrs = TRUE,
  net_attrs = TRUE)

Arguments

x
igraph (igraph::graph), network::network, or tidygraph::tbl_graph.

...
Named arguments passed on to other methods. Currently edg_attrs, vrt_attrs, and net_attrs.

edg_attrs
logical, defaults TRUE. Whether to retain x's edge attributes.

vrt_attrs
logical, defaults TRUE. Whether to retain x's vertex attributes.

net_attrs
logical, defaults TRUE. Whether to retain xs network/graph-level attributes.

Details

sna_net objects contain the following elements:

graph_attributes

graph_attributes describe external data fields used to annotate objects, such as graph names or citation information.

metadata

metadata describe the graph itself and include:

n_vertices: integer counting the number of vertices in the graph.

n_edges: integer counting the number of edges in the graph.

is_directed: logical indicating whether edges are intended to be directed.

is_bipartite: logical indicating whether the graph is biparite

n_actors: integer counting the number of vertices belonging to a bipartite graph's "actor" mode.

any_multiplex: logical indicating whether the graph includes any parallel edges.

any_loops: logical indicating whether the graph contains loop edges.

any_isolates: logical indicating whether the graph contains any isolated vertices.

edges

edges are contained in either a matrix or data.frame where the first column (.ego) always contains an edge's source and the second column (.alter) always contains an edge's target.

.ego and .alter are stored as type integer.

If the edge data of an sna_net object only include .egos and .alters, edges are stored in a two-column matrix.

If edge attribute data are present, edges are stored in a data.frame where each additional column contains a given edge attribute.

vertices

vertices are contained in a data.frame where the first column always contains vertex .names and subsequent columns contain any vertex attributes present.

If an sna_net contains a bipartite graph, the second column will always be .actor, a logical vector indicating whether each vertex is intended to belong to the "actor" mode.