Davis' Southern Women (Deep South)

southern_women is an adaptation of the Pajek dataset "DeepSouth.paj".

southern_women

Format

Object Class: igraph

directed: FALSE

bipartite: TRUE

actors: 18

non-actors: 14

multiplex: FALSE

vertices: 32 ("novices")

edges: 89 ("affect relations")

loops: FALSE

Vertex Attributes:

name

character representing each person's name.

type/.actor

logical indicating whether each vertex belong's to the networks actor mode.

Edge Attributes:

event_date

character indicating the month and day that the event occurred.

event_month

numeric indicating the month that the event occurred.

event_day

numeric indicating the day that the vent occurred.

Source

http://vlado.fmf.uni-lj.si/pub/networks/data/GBM/davis.htm

Details

From the UC Irvine Network Data Repository documentation:

This is a data set of 18 women observed over a nine-month period. During that period, various subsets of these women met in a series of 14 informal social events. The data recored which women met for which events.

This data is originally from Davis, Gardner and Gardner (1941) via UCINET. This documentation is taken from Freeman (2003) in his usual lucid description. See the reference to the paper below:

In the 1930s, five ethnographers, Allison Davis, Elizabeth Stubbs Davis, Burleigh B. Gardner, Mary R. Gardner and J. G. St. Clair Drake, collected data on stratification in Natchez, Mississippi (Warner, 1988, p. 93). They produced the book cited below (Davis, et al., 1941) that reported a comparative study of social class in black and in white society. One element of this work involved examining the correspondence between people's social class levels and their patterns of informal interaction. Davis et. al was concerned with the issue of how much the informal contacts made by individuals were established solely (or primarily) with others at approximately their own class levels. To address this question the authors collected data on social events and examined people's patterns of informal contacts.

In particular, they collected systematic data on the social activities of 18 women whom they observed over a nine-month period. During that period, various subsets of these women had met in a series of 14 informal social events. The participation of women in events was uncovered using “interviews, the records of participant observers, guest lists, and the newspapers” (Davis et al., p. 149). Homans (1950, p. 82), who presumably had been in touch with the research team, reported that the data reflect joint activities like, “a day's work behind the counter of a store, a meeting of a women's club, a church supper, a card party, a supper party, a meeting of the Parent-Teacher Association, etc.”

This data set has several interesting properties. It is small and manageable. It embodies a relatively simple structural pattern, one in which, according to Davis et al., the women seemed to organize themselves into two more or less distinct groups. Moreover, they reported that the positions - core and peripheral - of the members of these groups could also be determined in terms of the ways in which different women had been involved in group activities.

At the same time, the Souther Women data set is complicated enough that some of the details of its patterning are less than obvious. As Homans (1950, p. 84) put it, “The pattern is frayed at the edges.“ And, finally, this data set comes to us in a two-mode “woman by event” form. Thus, it provides an opportunity to explore methods designed for direct application to two-mode data. But at the same time, it can easily be transformed into two one-mode matrices (woman by woman or event by event) that can be examined using tools for one-mode analysis.

Because of these properties, this Souther Women data set has become something of a touchstone for comparing analytic methods in social network analysis. Davis, Gardner and Gardner presented an intuitive interpretation of the data, based in part on their ethnographic experience in the community. Then the Souther Women data set was picked up by Homans (1950) who provided an alternative intuitive interpretation. In 1972, Phillips and Conviser used an analytic tool, based on information theory, that provided a systematic way to reexamine the Souther Women data. Since then, this data set has been analyzed again and again. It reappears whenever any network analyst wants to explore the utility of some new tool for analyzing data.

References

Breiger R. (1974). The duality of persons and groups. Social Forces, 53, 181-190.

Davis, A et al. (1941). Deep South. Chicago: University of Chicago Press.

Linton C. Freeman (2003). Finding Social Groups: A Meta-Analysis of the Southern Women Data, In Ronald Breiger, Kathleen Carley and Philippa Pattison, eds. Dynamic Social Network Modeling and Analysis. Washington: The National Academies Press.

Examples

library(snatools)

southern_women

#> IGRAPH 024dfe5 UN-B 32 89 -- 
#> + attr: .vrt_id (v/n), name (v/c), .actor (v/l), type (v/l), .edg_id
#> | (e/n), event_date (e/c), event_month (e/n), event_day (e/n)
#> + edges from 024dfe5 (vertex names):
#>  [1] Evelyn   --E01 Evelyn   --E02 Evelyn   --E03 Evelyn   --E04 Evelyn   --E05
#>  [6] Evelyn   --E06 Evelyn   --E08 Evelyn   --E09 Laura    --E01 Laura    --E02
#> [11] Laura    --E03 Laura    --E05 Laura    --E06 Laura    --E07 Laura    --E08
#> [16] Theresa  --E02 Theresa  --E03 Theresa  --E04 Theresa  --E05 Theresa  --E06
#> [21] Theresa  --E07 Theresa  --E08 Theresa  --E09 Brenda   --E01 Brenda   --E03
#> [26] Brenda   --E04 Brenda   --E05 Brenda   --E06 Brenda   --E07 Brenda   --E08
#> [31] Charlotte--E03 Charlotte--E04 Charlotte--E05 Charlotte--E07 Frances  --E03
#> + ... omitted several edges

as_network(southern_women)

#>  Network attributes:
#>   vertices = 32 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = 18 
#>   total edges= 89 
#>     missing edges= 0 
#>     non-missing edges= 89 
#> 
#>  Vertex attribute names: 
#>     .actor .vrt_id vertex.names 
#> 
#>  Edge attribute names: 
#>     .edg_id event_date event_day event_month