Filter vertices by their attributes.

Filter vertices by their attributes.

vrt_filter(x, ...)

vrt_filter_se(x, vrt_attr, .f, vrt_val = NULL)

# S3 method for bridge_net
vrt_filter(x, ...)

# S3 method for bridge_net
vrt_filter_se(x, vrt_attr, .f = NULL,
  vrt_val = NULL)

# S3 method for igraph
vrt_filter(x, ...)

# S3 method for igraph
vrt_filter_se(x, vrt_attr, .f = NULL, vrt_val = NULL)

# S3 method for network
vrt_filter(x, ...)

# S3 method for network
vrt_filter_se(x, vrt_attr, .f = NULL, vrt_val = NULL)

Arguments

x
A bridge_net, igraph, network, or tbl_graph.

...
logical conditions used to subset rows when said conditions evaluate to TRUE.

vrt_attr
character specifying edge attribute to target. vrt_filter_se() only.

.f
function (default: NULL) specifying the predicate function to apply to vrt_attr. vrt_filter_se() only.

vrt_val
(default: NULL) Value to use as second argument to .f (if applicable).

Value

x, but with vertices filtered by ... for vrt_filter() or edge_attr, .f, and/or vrt_val for vrt_filter_se().

Details

vrt_filter() uses non-standard evaluation to make syntax as convenient as possible, while vrt_filter_se() uses standard evaluation if preferred (or if vrt_filter() causes trouble).

Examples

library(snatools)

data(sampson_monastery)

ig <- sampson_monastery %>%
  as_igraph()

nw <- sampson_monastery %>%
  as_network()

sampson_monastery %>%
  vrt_filter(faction == "Outcasts", status == "Expelled")

#> # A directed, multiplex, 1-mode `bridge_net`.
#> # - Contains 0 loops and >0 isolates.
#> $edges # first 3 of 18
#>    .ego       .alter     weight relation time  positive_relation
#>    <.name>    <.name>    <dbl>  <chr>    <int> <lgl>            
#>  | Elias      Simplicius 1      liking   1     TRUE             
#>  | Elias      Simplicius 3      liking   2     TRUE             
#>  | Simplicius Elias      3      liking   2     TRUE             
#> $vertices # first 3 of 3
#>    .name      faction  cloisterville status  
#>    <.name>    <chr>    <lgl>         <chr>   
#>  | Basil      Outcasts  TRUE         Expelled
#>  | Elias      Outcasts FALSE         Expelled
#>  | Simplicius Outcasts FALSE         Expelled

sampson_monastery %>%
  vrt_filter_se("faction", `==`, "Outcasts") %>%
  vrt_filter_se("status", `==`, "Expelled")

#> # A directed, multiplex, 1-mode `bridge_net`.
#> # - Contains 0 loops and >0 isolates.
#> $edges # first 3 of 18
#>    .ego       .alter     weight relation time  positive_relation
#>    <.name>    <.name>    <dbl>  <chr>    <int> <lgl>            
#>  | Elias      Simplicius 1      liking   1     TRUE             
#>  | Elias      Simplicius 3      liking   2     TRUE             
#>  | Simplicius Elias      3      liking   2     TRUE             
#> $vertices # first 3 of 3
#>    .name      faction  cloisterville status  
#>    <.name>    <chr>    <lgl>         <chr>   
#>  | Basil      Outcasts  TRUE         Expelled
#>  | Elias      Outcasts FALSE         Expelled
#>  | Simplicius Outcasts FALSE         Expelled

ig %>%
  vrt_filter(cloisterville, faction == "Loyal")

#> IGRAPH 6741d76 DNW- 2 2 -- 
#> + attr: network_name (g/c), author (g/c), faction (v/c), cloisterville
#> | (v/l), status (v/c), name (v/c), weight (e/n), relation (e/c), time
#> | (e/n), positive_relation (e/l)
#> + edges from 6741d76 (vertex names):
#> [1] Bonaventure->Peter Bonaventure->Peter

ig %>%
  vrt_filter_se("cloisterville") %>%
  vrt_filter_se("faction", `==`, "Loyal")

#> IGRAPH 674511f DNW- 2 2 -- 
#> + attr: network_name (g/c), author (g/c), faction (v/c), cloisterville
#> | (v/l), status (v/c), name (v/c), weight (e/n), relation (e/c), time
#> | (e/n), positive_relation (e/l)
#> + edges from 674511f (vertex names):
#> [1] Bonaventure->Peter Bonaventure->Peter

nw %>%
  vrt_filter(status %in% c("Expelled", "Left Voluntarily"))

#>  Network attributes:
#>   vertices = 9 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = TRUE 
#>   bipartite = FALSE 
#>   network_name = Crisis in the Cloister 
#>   author = Samuel F. Sampson 
#>   total edges= 138 
#>     missing edges= 0 
#>     non-missing edges= 138 
#> 
#>  Vertex attribute names: 
#>     cloisterville faction status vertex.names 
#> 
#>  Edge attribute names: 
#>     positive_relation relation time weight 

nw %>%
  vrt_filter_se("status", `%in%`, c("Expelled", "Left Voluntarily"))

#>  Network attributes:
#>   vertices = 9 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = TRUE 
#>   bipartite = FALSE 
#>   network_name = Crisis in the Cloister 
#>   author = Samuel F. Sampson 
#>   total edges= 138 
#>     missing edges= 0 
#>     non-missing edges= 138 
#> 
#>  Vertex attribute names: 
#>     cloisterville faction status vertex.names 
#> 
#>  Edge attribute names: 
#>     positive_relation relation time weight