Conversion to network objects.

Conversion to network objects.

as_network(x, ...)

# S3 method for network
as_network(x, ...)

# S3 method for network_dataset
as_network(x)

# S3 method for igraph
as_network(x, actor_type = TRUE)

# S3 method for tbl_graph
as_network(x, ...)

Arguments

x
igraph (igraph::graph) or tidygraph::tbl_graph object.

...
Additional arguments passed on to other methods (currently actor_type).

actor_type
logical. For bipartite graphs, the vertex "type" specifying which vertices in x are to be treated as "actors" in returned network object. See Details.

Value

A network::network object.

Details

actor_type

By default, igraph represents bipartite graphs using a logical vertex attribute named type.

network represents bipartite graphs through a count of the vertices in the "actors" partition.

actor_type provides a way to specify which vertices of an igraph object should be represented as the "actors" in the resulting network object (as opposed to the vertices of the second mode, e.g. "events").

See also

Examples

library(snatools)

# `igraph` to `network` ===============================================================
## undirected =========================================================================
data("karate", package = "igraphdata")
karate

#> IGRAPH 4b458a1 UNW- 34 78 -- Zachary's karate club network
#> + attr: name (g/c), Citation (g/c), Author (g/c), Faction (v/n), name
#> | (v/c), label (v/c), color (v/n), weight (e/n)
#> + edges from 4b458a1 (vertex names):
#>  [1] Mr Hi  --Actor 2  Mr Hi  --Actor 3  Mr Hi  --Actor 4  Mr Hi  --Actor 5 
#>  [5] Mr Hi  --Actor 6  Mr Hi  --Actor 7  Mr Hi  --Actor 8  Mr Hi  --Actor 9 
#>  [9] Mr Hi  --Actor 11 Mr Hi  --Actor 12 Mr Hi  --Actor 13 Mr Hi  --Actor 14
#> [13] Mr Hi  --Actor 18 Mr Hi  --Actor 20 Mr Hi  --Actor 22 Mr Hi  --Actor 32
#> [17] Actor 2--Actor 3  Actor 2--Actor 4  Actor 2--Actor 8  Actor 2--Actor 14
#> [21] Actor 2--Actor 18 Actor 2--Actor 20 Actor 2--Actor 22 Actor 2--Actor 31
#> [25] Actor 3--Actor 4  Actor 3--Actor 8  Actor 3--Actor 9  Actor 3--Actor 10
#> + ... omitted several edges

as_network(karate)

#>  Network attributes:
#>   vertices = 34 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   multiple = FALSE 
#>   name = Zachary's karate club network 
#>   Citation = Wayne W. Zachary. An Information Flow Model for Conflict and Fission in Small Groups. Journal of Anthropological Research Vol. 33, No. 4 452-473 
#>   Author = Wayne W. Zachary 
#>   total edges= 78 
#>     missing edges= 0 
#>     non-missing edges= 78 
#> 
#>  Vertex attribute names: 
#>     color Faction label vertex.names 
#> 
#>  Edge attribute names: 
#>     weight 

## directed ===========================================================================
data("USairports", package = "igraphdata")

USairports

#> IGRAPH bf6202d DN-- 755 23473 -- US airports
#> + attr: name (g/c), name (v/c), City (v/c), Position (v/c), Carrier
#> | (e/c), Departures (e/n), Seats (e/n), Passengers (e/n), Aircraft
#> | (e/n), Distance (e/n)
#> + edges from bf6202d (vertex names):
#>  [1] BGR->JFK BGR->JFK BOS->EWR ANC->JFK JFK->ANC LAS->LAX MIA->JFK EWR->ANC
#>  [9] BJC->MIA MIA->BJC TEB->ANC JFK->LAX LAX->JFK LAX->SFO AEX->LAS BFI->SBA
#> [17] ELM->PIT GEG->SUN ICT->PBI LAS->LAX LAS->PBI LAS->SFO LAX->LAS PBI->AEX
#> [25] PBI->ICT PIT->VCT SFO->LAX VCT->DWH IAD->JFK ABE->CLT ABE->HPN AGS->CLT
#> [33] AGS->CLT AVL->CLT AVL->CLT AVP->CLT AVP->PHL BDL->CLT BHM->CLT BHM->CLT
#> [41] BNA->CLT BNA->CLT BNA->DCA BNA->PHL BTR->CLT BUF->CLT BUF->DCA BUF->PHL
#> + ... omitted several edges

as_network(USairports)

#>  Network attributes:
#>   vertices = 755 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   multiple = TRUE 
#>   name = US airports 
#>   total edges= 23473 
#>     missing edges= 0 
#>     non-missing edges= 23473 
#> 
#>  Vertex attribute names: 
#>     City Position vertex.names 
#> 
#>  Edge attribute names not shown 

## bipartite ==========================================================================
southern_women_matrix <- matrix(
  c(1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0,
    1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1,
    1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0,
    0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1,
    1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
    1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 1,
    1, 1, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0,
    0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0,
    0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0,
    1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0),
  nrow = 14L, ncol = 18L,
  dimnames = list(
    c("E1", "E2", "E3", "E4", "E5", "E6", "E7",
      "E8", "E9", "E10", "E11", "E12", "E13", "E14"),
    c("EVELYN", "LAURA", "THERESA", "BRENDA", "CHARLOTTE", "FRANCES",
      "ELEANOR", "PEARL", "RUTH", "VERNE", "MYRNA", "KATHERINE",
      "SYLVIA", "NORA", "HELEN", "DOROTHY", "OLIVIA", "FLORA"))
 )

southern_women_ig <- igraph::graph_from_incidence_matrix(southern_women_matrix)

southern_women_ig

#> IGRAPH 3610855 UN-B 32 93 -- 
#> + attr: type (v/l), name (v/c)
#> + edges from 3610855 (vertex names):
#>  [1] E1--EVELYN    E1--LAURA     E1--BRENDA    E2--EVELYN    E2--LAURA    
#>  [6] E2--THERESA   E3--EVELYN    E3--LAURA     E3--THERESA   E3--BRENDA   
#> [11] E3--CHARLOTTE E3--FRANCES   E4--EVELYN    E4--THERESA   E4--BRENDA   
#> [16] E4--CHARLOTTE E5--EVELYN    E5--LAURA     E5--THERESA   E5--BRENDA   
#> [21] E5--CHARLOTTE E5--FRANCES   E5--ELEANOR   E5--RUTH      E6--EVELYN   
#> [26] E6--LAURA     E6--THERESA   E6--BRENDA    E6--FRANCES   E6--ELEANOR  
#> [31] E6--PEARL     E6--NORA      E7--LAURA     E7--THERESA   E7--BRENDA   
#> [36] E7--CHARLOTTE E7--ELEANOR   E7--RUTH      E7--VERNE     E7--SYLVIA   
#> + ... omitted several edges

vrt_get_attrs(southern_women_ig)

#> # A tibble: 32 x 2
#>    type  name 
#>  * <lgl> <chr>
#>  1 FALSE E1   
#>  2 FALSE E2   
#>  3 FALSE E3   
#>  4 FALSE E4   
#>  5 FALSE E5   
#>  6 FALSE E6   
#>  7 FALSE E7   
#>  8 FALSE E8   
#>  9 FALSE E9   
#> 10 FALSE E10  
#> # ... with 22 more rows

as_network(southern_women_ig)

#>  Network attributes:
#>   vertices = 32 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   multiple = FALSE 
#>   bipartite = 18 
#>   total edges= 93 
#>     missing edges= 0 
#>     non-missing edges= 93 
#> 
#>  Vertex attribute names: 
#>     is_actor vertex.names 
#> 
#> No edge attributes

## using `actor_type` to specify bipartite mapping ====================================

sw_transposed <- t(southern_women_matrix)

southern_women_with_actors_false <- igraph::graph_from_incidence_matrix(sw_transposed)

vrt_get_attrs(southern_women_with_actors_false)

#> # A tibble: 32 x 2
#>    type  name     
#>  * <lgl> <chr>    
#>  1 FALSE EVELYN   
#>  2 FALSE LAURA    
#>  3 FALSE THERESA  
#>  4 FALSE BRENDA   
#>  5 FALSE CHARLOTTE
#>  6 FALSE FRANCES  
#>  7 FALSE ELEANOR  
#>  8 FALSE PEARL    
#>  9 FALSE RUTH     
#> 10 FALSE VERNE    
#> # ... with 22 more rows

as_network(southern_women_with_actors_false, actor_type = FALSE)

#>  Network attributes:
#>   vertices = 32 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   multiple = FALSE 
#>   bipartite = 18 
#>   total edges= 93 
#>     missing edges= 0 
#>     non-missing edges= 93 
#> 
#>  Vertex attribute names: 
#>     is_actor vertex.names 
#> 
#> No edge attributes

# `tbl_graph` to `network` ============================================================
tidy_g <- tidygraph::create_notable("Zachary")

tidy_g

#> # A tbl_graph: 34 nodes and 78 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 34 x 0 (active)
#> #
#> # Edge Data: 78 x 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     1     3
#> 3     1     4
#> # ... with 75 more rows

as_network(tidy_g)

#>  Network attributes:
#>   vertices = 34 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   multiple = FALSE 
#>   name = Zachary 
#>   total edges= 78 
#>     missing edges= 0 
#>     non-missing edges= 78 
#> 
#> No vertex attributes
#> 
#> No edge attributes