Extract Vertex Attributes

Extract vertex attributes in a variety of formats consistent across igraph and network objects.

vrt_get_attrs(x, ...)

# S3 method for igraph
vrt_get_attrs(x)

# S3 method for network
vrt_get_attrs(x, ignore_na = TRUE)

vrt_get_attr(x, vrt_attr)

# S3 method for igraph
vrt_get_attr(x, vrt_attr)

# S3 method for network
vrt_get_attr(x, vrt_attr)

vrt_get_names(x)

# S3 method for igraph
vrt_get_names(x)

# S3 method for network
vrt_get_names(x)

vrt_get_attr_names(x, ...)

# S3 method for igraph
vrt_get_attr_names(x)

# S3 method for network
vrt_get_attr_names(x, ignore_na = TRUE)

Arguments

x
A graph object.

...
Arguments passed on to other methods. See below.

ignore_na
logical. Whether to ignore the "na" attribute of network::network objects.
Default: TRUE

vrt_attr
character. Name of target attribute.

Value

A named list, vector, or tibble::tibble. See Details.

Details

vrt_get_attrs() extracts all vertex attributes as a named list.

vrt_get_attr() extracts the single vertex attribute specified by vrt_attr as a vector.

vrt_get_names() is a convenience wrapper around vrt_get_attr() to extract the names of vertices as a vector.

It assumes igraph objects use "name" and network objects use "vertex.names".

If a vertex attribute following this convention is not present, vrt_get_names() throws a warning and returns NULL.

vrt_get_attr_names() extract the names of vertex attributes as a character vector.

vrt_get_attr_names() differs from network::list.vertex.attributes(), which sorts attribute names alphabetically.

See also

Examples

library(snatools)

data("sampson_monastery")

(ig <- as_igraph(sampson_monastery))

#> IGRAPH 58a37cf DNW- 18 510 -- 
#> + attr: name (v/c), faction (v/c), cloisterville (v/l), status (v/c),
#> | weight (e/n), relation (e/c), time (e/n), positive_relation (e/l)
#> + edges from 58a37cf (vertex names):
#>  [1] Romauld    ->Ambrose     Romauld    ->Peter       Romauld    ->Albert     
#>  [4] Bonaventure->Romauld     Bonaventure->Victor      Bonaventure->Albert     
#>  [7] Ambrose    ->Romauld     Ambrose    ->Bonaventure Ambrose    ->Elias      
#> [10] Berthold   ->Peter       Berthold   ->Louis       Berthold   ->Gregory    
#> [13] Peter      ->Berthold    Peter      ->Hugh        Peter      ->Mark       
#> [16] Louis      ->Romauld     Louis      ->Berthold    Louis      ->John Bosco 
#> [19] Victor     ->Bonaventure Victor     ->Winfrid     Victor     ->Basil      
#> + ... omitted several edges

(nw <- as_network(sampson_monastery))

#>  Network attributes:
#>   vertices = 18 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   multiple = TRUE 
#>   total edges= 510 
#>     missing edges= 0 
#>     non-missing edges= 510 
#> 
#>  Vertex attribute names: 
#>     cloisterville faction status vertex.names 
#> 
#>  Edge attribute names: 
#>     positive_relation relation time weight 

vrt_get_attrs(ig)

#> # A tibble: 18 x 4
#>    name        faction     cloisterville status          
#>  * <chr>       <chr>       <lgl>         <chr>           
#>  1 Romauld     Waverers    FALSE         Remained        
#>  2 Bonaventure Loyal       TRUE          Remained        
#>  3 Ambrose     Loyal       FALSE         Remained        
#>  4 Berthold    Loyal       FALSE         Remained        
#>  5 Peter       Loyal       TRUE          Remained        
#>  6 Louis       Loyal       FALSE         Remained        
#>  7 Victor      Waverers    FALSE         Remained        
#>  8 Winfrid     Young Turks FALSE         Remained        
#>  9 John Bosco  Young Turks TRUE          Left Voluntarily
#> 10 Gregory     Young Turks TRUE          Expelled        
#> 11 Hugh        Young Turks FALSE         Left Voluntarily
#> 12 Boniface    Young Turks FALSE         Left Voluntarily
#> 13 Mark        Young Turks TRUE          Left Voluntarily
#> 14 Albert      Young Turks FALSE         Left Voluntarily
#> 15 Amand       Waverers    FALSE         Remained        
#> 16 Basil       Outcasts    TRUE          Expelled        
#> 17 Elias       Outcasts    FALSE         Expelled        
#> 18 Simplicius  Outcasts    FALSE         Expelled        

vrt_get_attrs(nw)

#> # A tibble: 18 x 4
#>    vertex.names faction     cloisterville status          
#>    <chr>        <chr>       <lgl>         <chr>           
#>  1 Romauld      Waverers    FALSE         Remained        
#>  2 Bonaventure  Loyal       TRUE          Remained        
#>  3 Ambrose      Loyal       FALSE         Remained        
#>  4 Berthold     Loyal       FALSE         Remained        
#>  5 Peter        Loyal       TRUE          Remained        
#>  6 Louis        Loyal       FALSE         Remained        
#>  7 Victor       Waverers    FALSE         Remained        
#>  8 Winfrid      Young Turks FALSE         Remained        
#>  9 John Bosco   Young Turks TRUE          Left Voluntarily
#> 10 Gregory      Young Turks TRUE          Expelled        
#> 11 Hugh         Young Turks FALSE         Left Voluntarily
#> 12 Boniface     Young Turks FALSE         Left Voluntarily
#> 13 Mark         Young Turks TRUE          Left Voluntarily
#> 14 Albert       Young Turks FALSE         Left Voluntarily
#> 15 Amand        Waverers    FALSE         Remained        
#> 16 Basil        Outcasts    TRUE          Expelled        
#> 17 Elias        Outcasts    FALSE         Expelled        
#> 18 Simplicius   Outcasts    FALSE         Expelled        

vrt_get_attr(ig, vrt_attr = "faction")

#>  [1] "Waverers"    "Loyal"       "Loyal"       "Loyal"       "Loyal"      
#>  [6] "Loyal"       "Waverers"    "Young Turks" "Young Turks" "Young Turks"
#> [11] "Young Turks" "Young Turks" "Young Turks" "Young Turks" "Waverers"   
#> [16] "Outcasts"    "Outcasts"    "Outcasts"   

vrt_get_attr(nw, vrt_attr = "status")

#>  [1] "Remained"         "Remained"         "Remained"         "Remained"        
#>  [5] "Remained"         "Remained"         "Remained"         "Remained"        
#>  [9] "Left Voluntarily" "Expelled"         "Left Voluntarily" "Left Voluntarily"
#> [13] "Left Voluntarily" "Left Voluntarily" "Remained"         "Expelled"        
#> [17] "Expelled"         "Expelled"        

vrt_get_names(ig)

#>  [1] "Romauld"     "Bonaventure" "Ambrose"     "Berthold"    "Peter"      
#>  [6] "Louis"       "Victor"      "Winfrid"     "John Bosco"  "Gregory"    
#> [11] "Hugh"        "Boniface"    "Mark"        "Albert"      "Amand"      
#> [16] "Basil"       "Elias"       "Simplicius" 

vrt_get_names(nw)

#>  [1] "Romauld"     "Bonaventure" "Ambrose"     "Berthold"    "Peter"      
#>  [6] "Louis"       "Victor"      "Winfrid"     "John Bosco"  "Gregory"    
#> [11] "Hugh"        "Boniface"    "Mark"        "Albert"      "Amand"      
#> [16] "Basil"       "Elias"       "Simplicius" 

vrt_get_attr_names(ig)

#> [1] "name"          "faction"       "cloisterville" "status"       

vrt_get_attr_names(nw)

#> [1] "vertex.names"  "faction"       "cloisterville" "status"